Chapter 6 - Hypno Nerd
Taking the screens from chapter 5 we are now going to implement a
second screen that allows the user to set a time and date so that they
can be reminded to get hypnotised. This version see the use of a
UIViewController rather than explicitly setting views in the App
Delegate.